+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Next version
-- FTP030 will report an import of `from __future__ import annotations` if python 3.14 is used
+- FTP030 will report an import of `from __future__ import annotations` if python 3.14 onwards is used
 
 ## 25.5.22.0
 - Support async for loops for all checks working on for loops