+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next version
+- Improve FTP097 to also consider `StrEnum` and `IntEnum`
 
 ## 25.5.22.0
 - Support async for loops for all checks working on for loops