+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next version
 - Improve FTP097 to also consider `StrEnum` and `IntEnum`
 - FTP030 will report an import of `from __future__ import annotations` if python 3.14 onwards is used
+- FTP003 and FTP007 will now consider all usages of the functions not only calls
 
 ## 25.5.22.0
 - Support async for loops for all checks working on for loops